Lab 3 : type casted to float

+++ src/Lab/Lab3/P1/shape.java
@@ -1,5 +1,5 @@
 package Lab.Lab3.P1;
 
 public abstract class shape {
-    public abstract double area();
+    public abstract float area();
 }
